tests: move home_directory fixture to conftest
Will be useful in other tests
